Move session tests into separate module
Ref. #398
