+++ tests/test_sessionfile.py
@@ -89,202 +89,6 @@ def test_missing_states() -> None:
     )
 
 
-def test_empty_states() -> None:
-    with pytest.raises(
-        RecordFluxError,
-        match=(
-            "^"
-            "model: error: empty states\n"
-            'model: error: initial state "START" does not exist in "session"\n'
-            'model: error: final state "END" does not exist in "session"'
-            "$"
-        ),
-    ):
-        Session(
-            identifier="session", initial=ID("START"), final=ID("END"), states=[], declarations=[],
-        )
-
-
-def test_invalid_initial() -> None:
-    with pytest.raises(
-        RecordFluxError,
-        match=(
-            "^"
-            'model: error: initial state "NONEXISTENT" does not exist in "session"\n'
-            "model: error: unreachable states START"
-            "$"
-        ),
-    ):
-        Session(
-            identifier="session",
-            initial=ID("NONEXISTENT"),
-            final=ID("END"),
-            states=[
-                State(name=ID("START"), transitions=[Transition(target=ID("END"))]),
-                State(name=ID("END")),
-            ],
-            declarations=[],
-        )
-
-
-def test_invalid_final() -> None:
-    with pytest.raises(
-        RecordFluxError,
-        match=(
-            "^"
-            'model: error: final state "NONEXISTENT" does not exist in "session"\n'
-            "model: error: detached states END"
-            "$"
-        ),
-    ):
-        Session(
-            identifier="session",
-            initial=ID("START"),
-            final=ID("NONEXISTENT"),
-            states=[
-                State(name=ID("START"), transitions=[Transition(target=ID("END"))]),
-                State(name=ID("END")),
-            ],
-            declarations=[],
-        )
-
-
-def test_invalid_target_state() -> None:
-    with pytest.raises(
-        RecordFluxError,
-        match='^model: error: transition from state "START" to non-existent'
-        ' state "NONEXISTENT" in "session"$',
-    ):
-        Session(
-            identifier="session",
-            initial=ID("START"),
-            final=ID("END"),
-            states=[
-                State(
-                    name=ID("START"),
-                    transitions=[
-                        Transition(target=ID("NONEXISTENT")),
-                        Transition(target=ID("END")),
-                    ],
-                ),
-                State(name=ID("END")),
-            ],
-            declarations=[],
-        )
-
-
-def test_duplicate_state() -> None:
-    with pytest.raises(RecordFluxError, match="^model: error: duplicate states: START$"):
-        Session(
-            identifier="session",
-            initial=ID("START"),
-            final=ID("END"),
-            states=[
-                State(name=ID("START"), transitions=[Transition(target=ID("END"))]),
-                State(name=ID("START"), transitions=[Transition(target=ID("END"))]),
-                State(name=ID("END")),
-            ],
-            declarations=[],
-        )
-
-
-def test_multiple_duplicate_states() -> None:
-    with pytest.raises(
-        RecordFluxError, match=("^model: error: duplicate states: BAR, FOO, START$")
-    ):
-        Session(
-            identifier="session",
-            initial=ID("START"),
-            final=ID("END"),
-            states=[
-                State(name=ID("START"), transitions=[Transition(target=ID("FOO"))]),
-                State(name=ID("START"), transitions=[Transition(target=ID("FOO"))]),
-                State(name=ID("FOO"), transitions=[Transition(target=ID("BAR"))]),
-                State(name=ID("BAR"), transitions=[Transition(target=ID("END"))]),
-                State(name=ID("FOO"), transitions=[Transition(target=ID("BAR"))]),
-                State(name=ID("BAR"), transitions=[Transition(target=ID("END"))]),
-                State(name=ID("END")),
-            ],
-            declarations=[],
-        )
-
-
-def test_unreachable_state() -> None:
-    with pytest.raises(RecordFluxError, match="^model: error: unreachable states UNREACHABLE$"):
-        Session(
-            identifier="session",
-            initial=ID("START"),
-            final=ID("END"),
-            states=[
-                State(name=ID("START"), transitions=[Transition(target=ID("END"))]),
-                State(name=ID("UNREACHABLE"), transitions=[Transition(target=ID("END"))],),
-                State(name=ID("END")),
-            ],
-            declarations=[],
-        )
-
-
-def test_multiple_unreachable_states() -> None:
-    with pytest.raises(
-        RecordFluxError, match="^model: error: unreachable states UNREACHABLE1, UNREACHABLE2$"
-    ):
-        Session(
-            identifier="session",
-            initial=ID("START"),
-            final=ID("END"),
-            states=[
-                State(name=ID("START"), transitions=[Transition(target=ID("END"))]),
-                State(name=ID("UNREACHABLE1"), transitions=[Transition(target=ID("END"))],),
-                State(name=ID("UNREACHABLE2"), transitions=[Transition(target=ID("END"))],),
-                State(name=ID("END")),
-            ],
-            declarations=[],
-        )
-
-
-def test_detached_state() -> None:
-    with pytest.raises(RecordFluxError, match="^model: error: detached states DETACHED$"):
-        Session(
-            identifier="session",
-            initial=ID("START"),
-            final=ID("END"),
-            states=[
-                State(
-                    name=ID("START"),
-                    transitions=[Transition(target=ID("END")), Transition(target=ID("DETACHED"))],
-                ),
-                State(name=ID("DETACHED")),
-                State(name=ID("END")),
-            ],
-            declarations=[],
-        )
-
-
-def test_multiple_detached_states() -> None:
-    with pytest.raises(
-        RecordFluxError, match="^model: error: detached states DETACHED1, DETACHED2$"
-    ):
-        Session(
-            identifier="session",
-            initial=ID("START"),
-            final=ID("END"),
-            states=[
-                State(
-                    name=ID("START"),
-                    transitions=[
-                        Transition(target=ID("END")),
-                        Transition(target=ID("DETACHED1")),
-                        Transition(target=ID("DETACHED2")),
-                    ],
-                ),
-                State(name=ID("DETACHED1")),
-                State(name=ID("DETACHED2")),
-                State(name=ID("END")),
-            ],
-            declarations=[],
-        )
-
-
 def test_session_with_conditions() -> None:
     f = SessionFile()
     f.parse_string(
